Add ADC hardware abstraction layer
This commit add the abstraction layer for the ADC peripheral.

The user sets the configuration of the group, provide a buffer to store
the conversion results and a callback for when the group conversion
finishes.

In this version, the user should trigger manually the conversion using
adc_start(...). In the future external trigger, DMA and continuous
conversion will be implemented.